requirements.txt
Django>=3.2.8
psycopg2-binary==2.9.1
djangorestframework>=3.12.4
markdown>=3.3.4
django-filter>=2.4.0
paho-mqtt>=1.6.1
django-extra-views
django-crispy-forms